requirements.txt
names~=0.3.0
Pillow~=9.1.0
PyQt5~=5.15.4
fastapi~=0.78.0
uvicorn~=0.17.6